Clean up test app
